backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
black==23.12.1
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
fastapi==0.115.6
flatbuffers==24.3.25
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
protobuf==5.29.2
pydantic==2.10.4
pydantic_core==2.27.2
PyMatting==1.1.13
pytest==7.4.4
python-multipart==0.0.20
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.14.1
setuptools==75.6.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tifffile==2024.12.12
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
